% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_team_season_link.R
\name{build_team_season_link}
\alias{build_team_season_link}
\title{Title}
\usage{
build_team_season_link(html_link, season_code)
}
\arguments{
\item{season_code}{}
}
\description{
Title
}
\examples{

(season_code_2023 <- get_season_code(2023))

build_team_season_link(html_link = "https://tf.tfrrs.org/teams/tf/MI_college_m_Michigan.html",
                      season_code = season_code_2023)
}
