% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athlete_id.R
\name{get_athlete_id}
\alias{get_athlete_id}
\title{Get Athlete ID}
\usage{
get_athlete_id(html_table)
}
\arguments{
\item{html_table}{html_nodes, a table of results}
}
\value{
character, the unique href with id for the athlete
}
\description{
Get Athlete ID
}
\details{
This function is used to get the athlete ID from the results table. This is used to get the unique link to the athlete's page on TFRRS. It can successfully traverse the different table types for solo events and relays.
}
\examples{
"https://tf.tfrrs.org/all_performances/MI_college_m_Michigan.html?list_hnd=4541&season_hnd=645" \%>\%
rvest::read_html() \%>\%
rvest::html_nodes("table") \%>\%
magrittr::extract2(1) \%>\% # Just get the IDs for the first table in this example.
get_athlete_id()
}
