% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conference_links.R
\name{get_conference_links}
\alias{get_conference_links}
\title{Get Conference Links}
\usage{
get_conference_links(html_file)
}
\arguments{
\item{html_file}{}
}
\description{
Get Conference Links
}
\examples{

get_conference_links("ncaa_track_d1_html")
}
