% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_code.R
\name{get_season_code}
\alias{get_season_code}
\title{Get Season Code}
\usage{
get_season_code(year)
}
\arguments{
\item{year}{the year for which you want the season code}
}
\value{
numberic vector of length 1
}
\description{
See the `season_codes` data frame from this package. This function matches the year to the season code.
}
\examples{
get_season_code(2023)
}
