% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_links_from_conference.R
\name{get_team_links_from_conference}
\alias{get_team_links_from_conference}
\title{Get Team Links from Conference Link}
\usage{
get_team_links_from_conference(html_file)
}
\arguments{
\item{html_file}{the html file of a conference page to be parsed}
}
\value{
a character vector of unique team links
}
\description{
Takes a conference page and returns the unique site links for all teams in that conference
}
\examples{

"https://tf.tfrrs.org/lists/4759/Big_Ten_Outdoor_Performance_List?gender=m" \%>\%
rvest::read_html() \%>\%
get_team_links_from_conference()

}
