% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_yearly_results.R
\name{get_team_yearly_results}
\alias{get_team_yearly_results}
\title{Get Team Yearly Results}
\usage{
get_team_yearly_results(team_link, year)
}
\arguments{
\item{year}{}
}
\description{
Get the results tables for a team in a given year
}
\examples{
# See results for the University of Michigan in the 2021 season
get_team_yearly_results("https://tf.tfrrs.org/teams/tf/MI_college_m_Michigan.html", 2021)

}
