% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_year_from_season_code.R
\name{get_year_from_season_code}
\alias{get_year_from_season_code}
\title{Get Year from Season Code}
\usage{
get_year_from_season_code(season_code)
}
\arguments{
\item{season_code}{character or number of the}
}
\value{
numeric or character value representing the year
}
\description{
Convert the season_code value to the year/season of the event. See the `list_hnd` variable in the `season_codes` data object.
}
\examples{

get_year_from_season_code(600)

}
