% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncaatrackr_data.R
\name{load_ncaatrackr_data}
\alias{load_ncaatrackr_data}
\title{Load ncaatrackr Data}
\usage{
load_ncaatrackr_data()
}
\value{
a list of html files containing data from ncaatrackr
}
\description{
Loads in the html files stored in the package `extdata` folder
}
\examples{

load_ncaatrackr_data()

}
