% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaatrackr-package.R
\docType{package}
\name{ncaatrackr-package}
\alias{ncaatrackr}
\alias{ncaatrackr-package}
\title{ncaatrackr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
