% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_athlete_id.R
\name{parse_athlete_id}
\alias{parse_athlete_id}
\title{Parse Athlete ID}
\usage{
parse_athlete_id(athlete_link)
}
\arguments{
\item{athlete_link}{}
}
\value{
character string of numbers between 4 and 10 characters long
}
\description{
Takes an athlete link and returns just the unique ID number from the link
}
\examples{

parse_athlete_id("https://tf.tfrrs.org/athletes/2157720/Florida_State/")

}
