% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sex.R
\name{parse_sex}
\alias{parse_sex}
\title{Parse Sex}
\usage{
parse_sex(team_stub)
}
\arguments{
\item{team_stub}{}
}
\value{
character vector of length 1 stating either Male or Female
}
\description{
Parse the sex information from a link stub
}
\examples{

parse_sex("FL_college_f_Florida_State")

}
