% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_team_link.R
\name{parse_team_name}
\alias{parse_team_name}
\title{Parse Team Link}
\usage{
parse_team_name(team_link)
}
\arguments{
\item{team_link}{character URL link for a given team}
}
\value{
character string
}
\description{
Parse Team Link
}
\examples{

parse_team_name("https://tf.tfrrs.org/teams/tf/FL_college_f_Florida_State.html")

}
