% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_year_from_link.R
\name{parse_year_from_link}
\alias{parse_year_from_link}
\title{Get Year from Link}
\usage{
parse_year_from_link(team_season_link)
}
\arguments{
\item{team_season_link}{character URL link}
}
\value{
character or numeric value for the year
}
\description{
Parses the URL included in the `team_season_link` variable to find the year associated with the `list_hnd` value.
}
\examples{

"https://tf.tfrrs.org/all_performances/FL_college_f_Florida_State.html?list_hnd=600" \%>\%
parse_year_from_link()

}
