% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{season_codes}
\alias{season_codes}
\title{Season Codes}
\format{
## `season_codes`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{outdoor_season}{Season for track}
  \item{list_hnd}{Associated code for the season}
}
}
\usage{
season_codes
}
\description{
Codes that match the yearly season for track and field. This is called internally by some functions, and generally does not have a use for end-users.
}
\keyword{datasets}
