% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conference_links.R
\name{get_conference_links}
\alias{get_conference_links}
\title{Get Conference Links}
\usage{
get_conference_links(html_file)
}
\arguments{
\item{html_file}{The html file containing the conference links.}
}
\value{
vector of html links
}
\description{
Get Conference Links
}
\examples{
ncaatrackr::load_ncaatrackr_data()[["tfrrs_outdoor_lists_ncaad1.html"]] \%>\%
get_conference_links()

}
