% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_code.R
\name{get_season_code}
\alias{get_season_code}
\title{Get Season Code}
\usage{
get_season_code(year)
}
\arguments{
\item{year}{}
}
\value{
vector of length 1
}
\description{
See the `season_code` data frame from this package. This function matches the year to the season code.
}
