% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_link_stub.R
\name{get_team_link_stub}
\alias{get_team_link_stub}
\title{Get Team Link Stub}
\usage{
get_team_link_stub(team_link)
}
\arguments{
\item{team_link}{The base link to the team's page on TFRRS.}
}
\value{
character, the stub needed for building the team-season link.
}
\description{
Takes the base link to the team's page on TFRRS and returns the stub needed for building the team-season link.
}
\examples{

get_team_link_stub("https://tf.tfrrs.org/teams/tf/MI_college_m_Michigan.html")

}
