% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_links.R
\name{get_team_links_from_conference}
\alias{get_team_links_from_conference}
\title{Get Team Links from Conference Link}
\usage{
get_team_links_from_conference(html_file)
}
\arguments{
\item{html_file}{}
}
\description{
Takes a conference link and returns the site links for all teams in that conference
}
